components/master/backend/requirements.txt
Flask
flask_cors
gunicorn
python-dotenv
boto3
numpy
pika